Refactored Types location
