+++ server/src/main/java/edu/colostate/cs/cs414/chessirecoders/JungleServer/server/ServerResponse.java
@@ -3,6 +3,7 @@ package edu.colostate.cs.cs414.chessirecoders.JungleServer.server;
 import com.esotericsoftware.kryo.Kryo;
 import com.esotericsoftware.kryonet.EndPoint;
 import edu.colostate.cs.cs414.chessirecoders.JungleServer.data.GameHistory;
+import edu.colostate.cs.cs414.chessirecoders.JungleServer.data.Types;
 
 public class ServerResponse {
